Remove Plone favicon.ico
The SPI web site used Plone in the past and even though this hasn't been
true in years we still use the Plone favicon.ico.  Clearly there's no
reason to show the Plone logo for SPI and it's better not to have a
favicon.ico at all than display the logo of another project.