Remove login link
The "login" link is confusing.  It's a login to the member site, not
the SPI site.  The SPI site does not have a login.  And people only
need to access the member site to sign up (links are on the SPI web
site) or to vote (links are sent by email).