Use calendar years for annual reports
At the face-to-face meeting in New York in February 2016, we agreed that
the annual report should cover the calendar year.  Even though SPI's
elections and AGM are in July, SPI's fiscal year is the calendar year.
This brings the annual report in line with current practices and makes
it easier to prepare (the figures from the tax preparations for the
990 form can be used for the annual report rather than creating a new
treasurer report covering the period July to June).

Rename old annual reports to show that they cover two years and
update the 2015 report (formerly known as the 2016 report) to cover
the 2015 calendar year.