Build annual reports in Travis
Let's build all annual reports in Travis to ensure that pdflatex
continues to be able to build them.  We force a rebuild of all
annual reports with `touch *.tex`.  In a deployment environment,
we wouldn't want that because it forces the PDFs to change even
if the corresponding LaTeX source didn't change.