Associated Project HOWTO: Existing Contracts Undersigning
SPI is a public charity and cannot undersign (explicitly or implicitly) all contracts.
They should be disclosed prior to the formal application.

The HOWTO is not thinking in projects which already existed as foundations.
As a result, it is possible to a project join SPI with contracts not permissible to 501(c)(3) entities such as SPI, and cause a lot of headache.

Before the project join SPI the money might still not be under 501(c)(3) rules, which makes easier to find solutions and alternatives for these edge cases.
(Including, if relevant, alternatives to SPI itself which aren't public charities).