Add April 2025 meeting minutes
Note that the president's report was truncated in IRC, so this
should be extended to include a full copy when available. Also, the
registered guests list may be incomplete, I included one who
notified me in IRC that they attended but need to obtain a full list
from the former secretary. TODO lines have been added for both in
the interim.